tests: Refactored tests and added more tests, removed artifact tests
