+++ DecSm.Atom.Tests/BuildTests/Builds/MinimalAtomBuild.cs
@@ -1,0 +1,4 @@
+﻿namespace DecSm.Atom.Tests.BuildTests.Builds;
+
+[BuildDefinition]
+public sealed partial class MinimalAtomBuild : BuildDefinition, IBuildDefinition;
