+++ DecSm.Atom.Tests/DecSm.Atom.Tests.csproj
@@ -48,6 +48,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\DecSm.Atom\DecSm.Atom.csproj" />
+    <ProjectReference Include="..\DecSm.Atom.SourceGenerators\DecSm.Atom.SourceGenerators.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
   </ItemGroup>
 
 </Project>
