+++ DecSm.Atom.Tests/Utils/TestLogger.cs
@@ -1,0 +1,21 @@
+﻿namespace DecSm.Atom.Tests.Utils;
+
+public sealed class TestLogger : ILogger
+{
+    public StringBuilder LogContent { get; } = new();
+
+    public void Log<TState>(
+        LogLevel logLevel,
+        EventId eventId,
+        TState state,
+        Exception? exception,
+        Func<TState, Exception?, string> formatter) =>
+        LogContent.AppendLine(formatter(state, exception));
+
+    public bool IsEnabled(LogLevel logLevel) =>
+        true;
+
+    public IDisposable? BeginScope<TState>(TState state)
+        where TState : notnull =>
+        new DisposableAction(() => { });
+}
