+++ DecSm.Atom.Tests/_usings.cs
@@ -2,6 +2,7 @@
 global using System.IO.Abstractions;
 global using System.IO.Abstractions.TestingHelpers;
 global using System.Linq.Expressions;
+global using System.Text;
 global using DecSm.Atom.Args;
 global using DecSm.Atom.Artifacts;
 global using DecSm.Atom.Build;
@@ -14,6 +15,8 @@ global using DecSm.Atom.Params;
 global using DecSm.Atom.Paths;
 global using DecSm.Atom.Process;
 global using DecSm.Atom.Reports;
+global using DecSm.Atom.Tests.BuildTests.Builds;
+global using DecSm.Atom.Tests.Utils;
 global using DecSm.Atom.Util;
 global using DecSm.Atom.Variables;
 global using DecSm.Atom.Vaults;
@@ -28,3 +31,4 @@ global using NUnit.Framework;
 global using Shouldly;
 global using Spectre.Console;
 global using Spectre.Console.Testing;
+global using static DecSm.Atom.Tests.Utils.TestUtils;
